+++ api/static/js/leaflet-measure.js
@@ -1,471 +1,491 @@
-(function (factory, window) {
-    // define an AMD module that relies on 'leaflet'
-    if (typeof define === "function" && define.amd) {
-        define(["leaflet"], factory);
+import template from 'lodash/template';
 
-        // define a Common JS module that relies on 'leaflet'
-    } else if (typeof exports === "object") {
-        module.exports = factory(require("leaflet"));
+import units from './units';
+import calc from './calc';
+import * as dom from './dom';
+import { selectOne as $ } from './dom';
+import Symbology from './symbology';
+import { numberFormat } from './utils';
+
+import {
+  controlTemplate,
+  resultsTemplate,
+  pointPopupTemplate,
+  linePopupTemplate,
+  areaPopupTemplate
+} from './templates';
+
+const templateSettings = {
+  imports: { numberFormat },
+  interpolate: /{{([\s\S]+?)}}/g // mustache
+};
+const controlTemplateCompiled = template(controlTemplate, templateSettings);
+const resultsTemplateCompiled = template(resultsTemplate, templateSettings);
+const pointPopupTemplateCompiled = template(pointPopupTemplate, templateSettings);
+const linePopupTemplateCompiled = template(linePopupTemplate, templateSettings);
+const areaPopupTemplateCompiled = template(areaPopupTemplate, templateSettings);
+
+L.Control.Measure = L.Control.extend({
+  _className: 'leaflet-control-measure',
+  options: {
+    units: {},
+    position: 'topright',
+    primaryLengthUnit: 'feet',
+    secondaryLengthUnit: 'miles',
+    primaryAreaUnit: 'acres',
+    activeColor: '#ABE67E', // base color for map features while actively measuring
+    completedColor: '#C8F2BE', // base color for permenant features generated from completed measure
+    captureZIndex: 10000, // z-index of the marker used to capture measure events
+    popupOptions: {
+      // standard leaflet popup options http://leafletjs.com/reference-1.3.0.html#popup-option
+      className: 'leaflet-measure-resultpopup',
+      autoPanPadding: [10, 10]
+    }
+  },
+  initialize: function(options) {
+    L.setOptions(this, options);
+    const { activeColor, completedColor } = this.options;
+    this._symbols = new Symbology({ activeColor, completedColor });
+    this.options.units = L.extend({}, units, this.options.units);
+  },
+  onAdd: function(map) {
+    this._map = map;
+    this._latlngs = [];
+    this._initLayout();
+    map.on('click', this._collapse, this);
+    this._layer = L.layerGroup().addTo(map);
+    return this._container;
+  },
+  onRemove: function(map) {
+    map.off('click', this._collapse, this);
+    map.removeLayer(this._layer);
+  },
+  _initLayout: function() {
+    const className = this._className,
+      container = (this._container = L.DomUtil.create('div', `${className} leaflet-bar`));
+
+    container.innerHTML = controlTemplateCompiled({
+      model: {
+        className: className
+      }
+    });
+
+    // makes this work on IE touch devices by stopping it from firing a mouseout event when the touch is released
+    container.setAttribute('aria-haspopup', true);
+    L.DomEvent.disableClickPropagation(container);
+    L.DomEvent.disableScrollPropagation(container);
+
+    const $toggle = (this.$toggle = $('.js-toggle', container)); // collapsed content
+    this.$interaction = $('.js-interaction', container); // expanded content
+    const $start = $('.js-start', container); // start button
+    const $cancel = $('.js-cancel', container); // cancel button
+    const $finish = $('.js-finish', container); // finish button
+    this.$startPrompt = $('.js-startprompt', container); // full area with button to start measurment
+    this.$measuringPrompt = $('.js-measuringprompt', container); // full area with all stuff for active measurement
+    this.$startHelp = $('.js-starthelp', container); // "Start creating a measurement by adding points"
+    this.$results = $('.js-results', container); // div with coordinate, linear, area results
+    this.$measureTasks = $('.js-measuretasks', container); // active measure buttons container
+
+    this._collapse();
+    this._updateMeasureNotStarted();
+
+    if (!L.Browser.android) {
+      L.DomEvent.on(container, 'mouseenter', this._expand, this);
+      L.DomEvent.on(container, 'mouseleave', this._collapse, this);
+    }
+    L.DomEvent.on($toggle, 'click', L.DomEvent.stop);
+    if (L.Browser.touch) {
+      L.DomEvent.on($toggle, 'click', this._expand, this);
+    } else {
+      L.DomEvent.on($toggle, 'focus', this._expand, this);
+    }
+    L.DomEvent.on($start, 'click', L.DomEvent.stop);
+    L.DomEvent.on($start, 'click', this._startMeasure, this);
+    L.DomEvent.on($cancel, 'click', L.DomEvent.stop);
+    L.DomEvent.on($cancel, 'click', this._finishMeasure, this);
+    L.DomEvent.on($finish, 'click', L.DomEvent.stop);
+    L.DomEvent.on($finish, 'click', this._handleMeasureDoubleClick, this);
+  },
+  _expand: function() {
+    dom.hide(this.$toggle);
+    dom.show(this.$interaction);
+  },
+  _collapse: function() {
+    if (!this._locked) {
+      dom.hide(this.$interaction);
+      dom.show(this.$toggle);
+    }
+  },
+  // move between basic states:
+  // measure not started, started/in progress but no points added, in progress and with points
+  _updateMeasureNotStarted: function() {
+    dom.hide(this.$startHelp);
+    dom.hide(this.$results);
+    dom.hide(this.$measureTasks);
+    dom.hide(this.$measuringPrompt);
+    dom.show(this.$startPrompt);
+  },
+  _updateMeasureStartedNoPoints: function() {
+    dom.hide(this.$results);
+    dom.show(this.$startHelp);
+    dom.show(this.$measureTasks);
+    dom.hide(this.$startPrompt);
+    dom.show(this.$measuringPrompt);
+  },
+  _updateMeasureStartedWithPoints: function() {
+    dom.hide(this.$startHelp);
+    dom.show(this.$results);
+    dom.show(this.$measureTasks);
+    dom.hide(this.$startPrompt);
+    dom.show(this.$measuringPrompt);
+  },
+  // get state vars and interface ready for measure
+  _startMeasure: function() {
+    this._locked = true;
+    this._measureVertexes = L.featureGroup().addTo(this._layer);
+    this._captureMarker = L.marker(this._map.getCenter(), {
+      clickable: true,
+      zIndexOffset: this.options.captureZIndex,
+      opacity: 0
+    }).addTo(this._layer);
+    this._setCaptureMarkerIcon();
+
+    this._captureMarker
+      .on('mouseout', this._handleMapMouseOut, this)
+      .on('dblclick', this._handleMeasureDoubleClick, this)
+      .on('click', this._handleMeasureClick, this);
+
+    this._map
+      .on('mousemove', this._handleMeasureMove, this)
+      .on('mouseout', this._handleMapMouseOut, this)
+      .on('move', this._centerCaptureMarker, this)
+      .on('resize', this._setCaptureMarkerIcon, this);
+
+    L.DomEvent.on(this._container, 'mouseenter', this._handleMapMouseOut, this);
+
+    this._updateMeasureStartedNoPoints();
+
+    this._map.fire('measurestart', null, false);
+  },
+  // return to state with no measure in progress, undo `this._startMeasure`
+  _finishMeasure: function() {
+    const model = L.extend({}, this._resultsModel, { points: this._latlngs });
+
+    this._locked = false;
+
+    L.DomEvent.off(this._container, 'mouseover', this._handleMapMouseOut, this);
+
+    this._clearMeasure();
+
+    this._captureMarker
+      .off('mouseout', this._handleMapMouseOut, this)
+      .off('dblclick', this._handleMeasureDoubleClick, this)
+      .off('click', this._handleMeasureClick, this);
+
+    this._map
+      .off('mousemove', this._handleMeasureMove, this)
+      .off('mouseout', this._handleMapMouseOut, this)
+      .off('move', this._centerCaptureMarker, this)
+      .off('resize', this._setCaptureMarkerIcon, this);
+
+    this._layer.removeLayer(this._measureVertexes).removeLayer(this._captureMarker);
+    this._measureVertexes = null;
+
+    this._updateMeasureNotStarted();
+    this._collapse();
+
+    this._map.fire('measurefinish', model, false);
+  },
+  // clear all running measure data
+  _clearMeasure: function() {
+    this._latlngs = [];
+    this._resultsModel = null;
+    this._measureVertexes.clearLayers();
+    if (this._measureDrag) {
+      this._layer.removeLayer(this._measureDrag);
+    }
+    if (this._measureArea) {
+      this._layer.removeLayer(this._measureArea);
+    }
+    if (this._measureBoundary) {
+      this._layer.removeLayer(this._measureBoundary);
+    }
+    this._measureDrag = null;
+    this._measureArea = null;
+    this._measureBoundary = null;
+  },
+  // centers the event capture marker
+  _centerCaptureMarker: function() {
+    this._captureMarker.setLatLng(this._map.getCenter());
+  },
+  // set icon on the capture marker
+  _setCaptureMarkerIcon: function() {
+    this._captureMarker.setIcon(
+      L.divIcon({
+        iconSize: this._map.getSize().multiplyBy(2)
+      })
+    );
+  },
+  // format measurements to nice display string based on units in options
+  // `{ lengthDisplay: '100 Feet (0.02 Miles)', areaDisplay: ... }`
+  _getMeasurementDisplayStrings: function(measurement) {
+    const unitDefinitions = this.options.units;
+
+    return {
+      lengthDisplay: buildDisplay(
+        measurement.length,
+        this.options.primaryLengthUnit,
+        this.options.secondaryLengthUnit,
+        this.options.decPoint,
+        this.options.thousandsSep
+      ),
+      areaDisplay: buildDisplay(
+        measurement.area,
+        this.options.primaryAreaUnit,
+        this.options.secondaryAreaUnit,
+        this.options.decPoint,
+        this.options.thousandsSep
+      )
+    };
+
+    function buildDisplay(val, primaryUnit, secondaryUnit, decPoint, thousandsSep) {
+      if (primaryUnit && unitDefinitions[primaryUnit]) {
+        let display = formatMeasure(val, unitDefinitions[primaryUnit], decPoint, thousandsSep);
+        if (secondaryUnit && unitDefinitions[secondaryUnit]) {
+          const formatted = formatMeasure(
+            val,
+            unitDefinitions[secondaryUnit],
+            decPoint,
+            thousandsSep
+          );
+          display = `${display} (${formatted})`;
+        }
+        return display;
+      }
+      return formatMeasure(val, null, decPoint, thousandsSep);
     }
 
-    // attach your plugin to the global 'L' variable
-    if (typeof window !== "undefined" && window.L) {
-        factory(L);
+    function formatMeasure(val, unit, decPoint, thousandsSep) {
+      const unitDisplays = {
+        acres: __('acres'),
+        feet: __('feet'),
+        kilometers: __('kilometers'),
+        hectares: __('hectares'),
+        meters: __('meters'),
+        miles: __('miles'),
+        sqfeet: __('sqfeet'),
+        sqmeters: __('sqmeters'),
+        sqmiles: __('sqmiles')
+      };
+
+      const u = L.extend({ factor: 1, decimals: 0 }, unit);
+      const formattedNumber = numberFormat(
+        val * u.factor,
+        u.decimals,
+        decPoint || __('decPoint'),
+        thousandsSep || __('thousandsSep')
+      );
+      const label = unitDisplays[u.display] || u.display;
+      return [formattedNumber, label].join(' ');
     }
-})(function (L) {
-    L.Measure = {
-        linearMeasurement: "Distance measurement",
-        areaMeasurement: "Area measurement",
-        start: "Start",
-        meter: "m",
-        kilometer: "km",
-        squareMeter: "m²",
-        squareKilometers: "km²",
-    };
+  },
+  // update results area of dom with calced measure from `this._latlngs`
+  _updateResults: function() {
+    const calced = calc(this._latlngs);
+    const model = (this._resultsModel = L.extend(
+      {},
+      calced,
+      this._getMeasurementDisplayStrings(calced),
+      {
+        pointCount: this._latlngs.length
+      }
+    ));
+    this.$results.innerHTML = resultsTemplateCompiled({ model });
+  },
+  // mouse move handler while measure in progress
+  // adds floating measure marker under cursor
+  _handleMeasureMove: function(evt) {
+    if (!this._measureDrag) {
+      this._measureDrag = L.circleMarker(evt.latlng, this._symbols.getSymbol('measureDrag')).addTo(
+        this._layer
+      );
+    } else {
+      this._measureDrag.setLatLng(evt.latlng);
+    }
+    this._measureDrag.bringToFront();
+  },
+  // handler for both double click and clicking finish button
+  // do final calc and finish out current measure, clear dom and internal state, add permanent map features
+  _handleMeasureDoubleClick: function() {
+    const latlngs = this._latlngs;
+    let resultFeature, popupContent;
 
-    L.Control.Measure = L.Control.extend({
-        options: {
-            position: "topright",
-            title: "Measurement",
-            collapsed: true,
-            color: "#3d9dfc",
-        },
-        initialize: function (options) {
-            L.Util.setOptions(this, options);
-        },
-        onAdd: function (map) {
-            this._map = map;
-            this._container || this._initLayout();
-            return this._container;
-        },
-        _buildContainer: function () {
-            this._container = L.DomUtil.create("div", "leaflet-control-measure leaflet-bar leaflet-control");
+    this._finishMeasure();
 
-            this._contents = L.DomUtil.create("div", "leaflet-measure-contents", this._container);
+    if (!latlngs.length) {
+      return;
+    }
 
-            this._link = L.DomUtil.create("a", "leaflet-measure-toggle", this._container);
-            this._link.title = this.options.title || "Measurement";
-            this._link.href = "#";
+    if (latlngs.length > 2) {
+      latlngs.push(latlngs[0]); // close path to get full perimeter measurement for areas
+    }
 
-            if (this.options.title) {
-                var title = L.DomUtil.create("h5", "", this._contents);
-                title.innerText = this.options.title;
-            }
+    const calced = calc(latlngs);
 
-            this._buildItems();
-        },
-        _buildItems: function () {
-            var ele_ul = L.DomUtil.create("ul", "leaflet-measure-actions", this._contents);
-            var ele_li = L.DomUtil.create("li", "leaflet-measure-action", ele_ul);
-            var ele_link_line = L.DomUtil.create("a", "start", ele_li);
-            ele_link_line.innerText = L.Measure.linearMeasurement;
-            ele_link_line.href = "#";
-            L.DomEvent.disableClickPropagation(ele_link_line);
-            L.DomEvent.on(ele_link_line, "click", this._enableMeasureLine, this);
+    if (latlngs.length === 1) {
+      resultFeature = L.circleMarker(latlngs[0], this._symbols.getSymbol('resultPoint'));
+      popupContent = pointPopupTemplateCompiled({
+        model: calced
+      });
+    } else if (latlngs.length === 2) {
+      resultFeature = L.polyline(latlngs, this._symbols.getSymbol('resultLine'));
+      popupContent = linePopupTemplateCompiled({
+        model: L.extend({}, calced, this._getMeasurementDisplayStrings(calced))
+      });
+    } else {
+      resultFeature = L.polygon(latlngs, this._symbols.getSymbol('resultArea'));
+      popupContent = areaPopupTemplateCompiled({
+        model: L.extend({}, calced, this._getMeasurementDisplayStrings(calced))
+      });
+    }
 
-            ele_li = L.DomUtil.create("li", "leaflet-measure-action", ele_ul);
-            var ele_link_area = L.DomUtil.create("a", "leaflet-measure-action start", ele_li);
-            ele_link_area.innerText = L.Measure.areaMeasurement;
-            ele_link_area.href = "#";
-            L.DomEvent.disableClickPropagation(ele_link_area);
-            L.DomEvent.on(ele_link_area, "click", this._enableMeasureArea, this);
-        },
-        _initLayout: function () {
-            this._buildContainer();
-            L.DomEvent.disableClickPropagation(this._container);
-            L.DomEvent.disableScrollPropagation(this._container);
-            if (this.options.collapsed) {
-                L.DomEvent.on(
-                    this._container,
-                    {
-                        mouseenter: this._expand,
-                        mouseleave: this._collapse,
-                    },
-                    this
-                );
-            } else {
-                this._expand();
-            }
-        },
-        _enableMeasureLine: function (ev) {
-            L.DomEvent.stopPropagation(ev);
-            this._measureHandler = new L.MeasureAction(this._map, {
-                model: "distance",
-                color: this.options.color,
+    const popupContainer = L.DomUtil.create('div', '');
+    popupContainer.innerHTML = popupContent;
+
+    const zoomLink = $('.js-zoomto', popupContainer);
+    if (zoomLink) {
+      L.DomEvent.on(zoomLink, 'click', L.DomEvent.stop);
+      L.DomEvent.on(
+        zoomLink,
+        'click',
+        function() {
+          if (resultFeature.getBounds) {
+            this._map.fitBounds(resultFeature.getBounds(), {
+              padding: [20, 20],
+              maxZoom: 17
             });
-            this._measureHandler.enable();
+          } else if (resultFeature.getLatLng) {
+            this._map.panTo(resultFeature.getLatLng());
+          }
         },
-        _enableMeasureArea: function (ev) {
-            L.DomEvent.stopPropagation(ev);
-            this._measureHandler = new L.MeasureAction(this._map, {
-                model: "area",
-                color: this.options.color,
-            });
-            this._measureHandler.enable();
-        },
-        _expand: function () {
-            this._link.style.display = "none";
-            L.DomUtil.addClass(this._container, "leaflet-measure-expanded");
-            return this;
-        },
-        _collapse: function () {
-            this._link.style.display = "block";
-            L.DomUtil.removeClass(this._container, "leaflet-measure-expanded");
-            return this;
-        },
-    });
+        this
+      );
+    }
 
-    L.control.measure = L.control.Measure = function (options) {
-        return new L.Control.Measure(options);
-    };
+    const deleteLink = $('.js-deletemarkup', popupContainer);
+    if (deleteLink) {
+      L.DomEvent.on(deleteLink, 'click', L.DomEvent.stop);
+      L.DomEvent.on(
+        deleteLink,
+        'click',
+        function() {
+          // TODO. maybe remove any event handlers on zoom and delete buttons?
+          this._layer.removeLayer(resultFeature);
+        },
+        this
+      );
+    }
 
-    L.MeasureLable = L.Layer.extend({
-        options: {
-            offset: new L.Point(0, 30),
-            latlng: null,
-            content: "",
-            className: "",
-        },
-        initialize: function (options) {
-            L.Util.setOptions(this, options);
-        },
-        onAdd: function (map) {
-            this._map = map;
-            this._container || this._initLayout();
-            map._panes.popupPane.appendChild(this._container);
-            map.on("viewreset", this._updatePosition, this);
-            if (L.Browser.any3d) {
-                map.on("zoomanim", this._zoomAnimation, this);
-            }
-            this._update();
-        },
-        addTo: function (map) {
-            map.addLayer(this);
-            return this;
-        },
-        onRemove: function (map) {
-            map._panes.popupPane.removeChild(this._container);
-            map.off(
-                {
-                    viewreset: this._updatePosition,
-                    zoomanim: this._zoomAnimation,
-                },
-                this
-            );
-            this._map = null;
-        },
-        setLatLng: function (latlng) {
-            this.options.latlng = L.latLng(latlng);
-            this._updatePosition();
-            return this;
-        },
-        setContent: function (content) {
-            this.options.content = content;
-            this._updateContent();
-            return this;
-        },
-        _initLayout: function () {
-            this._container = L.DomUtil.create("div", this.options.className);
-            this._contentNode = L.DomUtil.create("div", "content", this._container);
-        },
-        _update: function () {
-            this._map && (this._updateContent(), this._updatePosition());
-        },
-        _updateContent: function () {
-            if (this.options.content) {
-                if (typeof this.options.content == "string") {
-                    this._contentNode.innerHTML = this.options.content;
-                } else {
-                    this._contentNode.innerHTML = "";
-                    this._contentNode.appendChild(this.options.content);
-                }
-            }
-        },
-        _updatePosition: function () {
-            var point = this._map.latLngToLayerPoint(this.options.latlng),
-                is3D = L.Browser.any3d,
-                offset = this.options.offset;
-            is3D && L.DomUtil.setPosition(this._container, point);
-            this._containerBottom = -offset.y - (is3D ? 0 : point.y);
-            this._containerLeft = offset.x + (is3D ? 0 : point.x);
-            this._container.style.bottom = this._containerBottom + "px";
-            this._container.style.left = this._containerLeft + "px";
-        },
-        _zoomAnimation: function (a) {
-            a = this._map._latLngToNewLayerPoint(this.options.latlng, a.zoom, a.center);
-            L.DomUtil.setPosition(this._container, a);
-        },
-        enableClose: function () {
-            this._closeButton = L.DomUtil.create("span", "close-measure", this._container);
-            this._closeButton.innerHTML =
-                '<svg class="icon" viewBox="0 0 40 40"><path stroke="#FF0000" stroke-width="3" d="M 10,10 L 30,30 M 30,10 L 10,30"/></svg>';
-            return this._closeButton;
-        },
-    });
+    resultFeature.addTo(this._layer);
+    resultFeature.bindPopup(popupContainer, this.options.popupOptions);
+    if (resultFeature.getBounds) {
+      resultFeature.openPopup(resultFeature.getBounds().getCenter());
+    } else if (resultFeature.getLatLng) {
+      resultFeature.openPopup(resultFeature.getLatLng());
+    }
+  },
+  // handle map click during ongoing measurement
+  // add new clicked point, update measure layers and results ui
+  _handleMeasureClick: function(evt) {
+    const latlng = this._map.mouseEventToLatLng(evt.originalEvent), // get actual latlng instead of the marker's latlng from originalEvent
+      lastClick = this._latlngs[this._latlngs.length - 1],
+      vertexSymbol = this._symbols.getSymbol('measureVertex');
 
-    L.MeasureAction = L.Handler.extend({
-        options: {
-            color: "#3d9dfc",
-            model: "distance", // area or distance
-        },
+    if (!lastClick || !latlng.equals(lastClick)) {
+      // skip if same point as last click, happens on `dblclick`
+      this._latlngs.push(latlng);
+      this._addMeasureArea(this._latlngs);
+      this._addMeasureBoundary(this._latlngs);
 
-        initialize: function (map, options) {
-            this._map = map;
-            this._map._measureHandler = this;
-            L.Util.setOptions(this, options);
-        },
-        setModel: function (model) {
-            this.options.model = model;
-            return this;
-        },
-        addHooks: function () {
-            this._activeMeasure();
-        },
-        removeHooks: function () {},
-        _activeMeasure: function () {
-            this._map._measureHandler._measurementStarted && this._map._measureHandler._finishMeasure();
-            this._measurementStarted ? this._finishMeasure() : this._enableMeasure();
-        },
-        _onMouseClick: function (event) {
-            var latlng = event.latlng || this._map.mouseEventToLatLng(event);
-            if (this._lastPoint && latlng.equals(this._lastPoint)) {
-                return;
-            }
-            if (this._trail.points.length > 0) {
-                var points = this._trail.points;
-                points.push(latlng);
-                var length = points.length;
-                this._totalDistance += this._getDistance(points[length - 2], points[length - 1]);
-                this._addMeasurePoint(latlng);
-                this._addMarker(latlng);
-                if (this.options.model !== "area") {
-                    this._addLable(latlng, this._getDistanceString(this._totalDistance), "leaflet-measure-lable");
-                }
-            } else {
-                this._totalDistance = 0;
-                this._addMeasurePoint(latlng);
-                this._addMarker(latlng);
-                if (this.options.model !== "area") {
-                    this._addLable(latlng, L.Measure.start, "leaflet-measure-lable");
-                }
-                this._trail.points.push(latlng);
-            }
-            this._lastPoint = latlng;
-            this._startMove = false;
-        },
-        _onMouseMove: function (event) {
-            var latlng = event.latlng;
-            if (this._trail.points.length > 0) {
-                if (this._startMove) {
-                    this._directPath.setLatLngs(this._trail.points.concat(latlng));
-                } else {
-                    this._directPath.setLatLngs([latlng]);
-                    this._startMove = true;
-                }
-            }
-        },
-        _enableMeasure: function () {
-            this._trail = {
-                overlays: [],
-                points: [],
-            };
-            var map = this._map;
-            L.DomUtil.addClass(map._container, "leaflet-measure-map");
-            map.contextMenu && map.contextMenu.disable();
-            this._measurementStarted = true;
-            map.on("click", this._onMouseClick, this);
-            map.on("dblclick contextmenu", this._finishMeasure, this);
-            map.doubleClickZoom.disable();
-            map.on("mousemove", this._onMouseMove, this);
-        },
-        _disableMeasure: function () {
-            var map = this._map;
-            L.DomUtil.removeClass(map.getContainer(), "leaflet-measure-map");
-            map.contextMenu && map.contextMenu.enable();
-            map.off("click", this._onMouseClick, this);
-            map.off("dblclick contextmenu", this._finishMeasure, this);
-            map.off("mousemove", this._onMouseMove, this);
-            map.doubleClickZoom.enable();
-            this._measurementStarted = this._startMove = false;
-            this.disable();
-        },
-        _finishMeasure: function (event) {
-            if (this._trail.points.length > 0) {
-                if (this._trail.points.length > 1) {
-                    if (!event || event.type === "contextmenu") {
-                        this._directPath.setLatLngs(this._trail.points);
-                    }
-                    if (this.options.model === "area") {
-                        this._addLable(
-                            this._lastPoint,
-                            this._getAreaString(this._trail.points),
-                            "leaflet-measure-lable",
-                            true
-                        );
-                    } else {
-                        this._addLable(
-                            this._lastPoint,
-                            this._getDistanceString(this._totalDistance),
-                            "leaflet-measure-lable",
-                            true
-                        );
-                    }
-                    this._directPath && this._map.removeLayer(this._directPath);
-                } else {
-                    this._clearOverlay.call(this);
-                }
-            }
-            this._disableMeasure();
-        },
-        _resetDirectPath: function (latlng) {
-            if (!this._directPath) {
-                if (this.options.model === "area") {
-                    this._directPath = new L.Polygon([latlng], {
-                        weight: 2,
-                        color: this.options.color,
-                        dashArray: [5, 5],
-                        fillOpacity: 0,
-                        interactive: false,
-                    });
-                } else {
-                    this._directPath = new L.Polyline([latlng], {
-                        weight: 2,
-                        color: this.options.color,
-                        dashArray: [5, 5],
-                        interactive: false,
-                    });
-                }
-                this._map.addLayer(this._directPath);
-                this._trail.overlays.push(this._directPath);
-            } else {
-                this._directPath.addLatLng(latlng);
-            }
-        },
-        _addMeasurePoint: function (latlng) {
-            if (!this._measurePath) {
-                if (this.options.model === "area") {
-                    this._measurePath = new L.Polygon([latlng], {
-                        weight: 2,
-                        color: this.options.color,
-                        fillColor: this.options.color,
-                        fillOpacity: 0.2,
-                        interactive: false,
-                    });
-                } else {
-                    this._measurePath = new L.Polyline([latlng], {
-                        weight: 2,
-                        color: this.options.color,
-                        interactive: false,
-                    });
-                }
-                this._map.addLayer(this._measurePath);
-                this._trail.overlays.push(this._measurePath);
-            } else {
-                this._measurePath.addLatLng(latlng);
-            }
-            this._resetDirectPath(latlng);
-        },
-        _addMarker: function (latLng) {
-            var marker = new L.CircleMarker(latLng, {
-                color: this.options.color,
-                opacity: 1,
-                weight: 1,
-                fillColor: "#FFFFFF",
-                fill: true,
-                fillOpacity: 1,
-                radius: 3,
-                interactive: false,
-            });
-            this._map.addLayer(marker);
-            this._trail.overlays.push(marker);
-        },
-        _addLable: function (latlng, content, className, ended) {
-            var lable = new L.MeasureLable({
-                latlng: latlng,
-                content: content,
-                className: className,
-            });
-            this._map.addLayer(lable);
-            this._trail.overlays.push(lable);
-            if (ended) {
-                var closeButton = lable.enableClose();
-                L.DomEvent.on(closeButton, "click", this._clearOverlay, this);
-            }
-        },
-        _clearOverlay: function () {
-            var i = 0,
-                overlays = this._trail.overlays,
-                length;
-            for (length = overlays.length; i < length; i++) {
-                this._map.removeLayer(overlays[i]);
-            }
-        },
-        toRadians: function (deg) {
-            return deg * (Math.PI / 180);
-        },
-        square: function (x) {
-            return Math.pow(x, 2);
-        },
-        _getDistanceString: function (distance) {
-            return distance < 1e3
-                ? this._numberFormat(distance, 0) + " " + L.Measure.meter
-                : this._numberFormat(distance / 1e3, 2) + " " + L.Measure.kilometer;
-        },
+      this._measureVertexes.eachLayer(function(layer) {
+        layer.setStyle(vertexSymbol);
+        // reset all vertexes to non-active class - only last vertex is active
+        // `layer.setStyle({ className: 'layer-measurevertex'})` doesn't work. https://github.com/leaflet/leaflet/issues/2662
+        // set attribute on path directly
+        if (layer._path) {
+          layer._path.setAttribute('class', vertexSymbol.className);
+        }
+      });
 
-        _getDistance: function (latlng1, latlng2) {
-            var earthRadius = 6378137; // radius of the earth in meter
-            var lat1 = this.toRadians(latlng1.lat);
-            var lat2 = this.toRadians(latlng2.lat);
-            var lat_dif = lat2 - lat1;
-            var lng_dif = this.toRadians(latlng2.lng - latlng1.lng);
-            var a =
-                this.square(Math.sin(lat_dif / 2)) +
-                Math.cos(lat1) * Math.cos(lat2) * this.square(Math.sin(lng_dif / 2));
-            return 2 * earthRadius * Math.asin(Math.sqrt(a));
-        },
-        _getAreaString: function (points) {
-            var a = Math.round(this._getArea(points));
-            return a < 1e6
-                ? this._numberFormat(a, 0) + " " + L.Measure.squareMeter
-                : this._numberFormat(a / 1e6, 2) + " " + L.Measure.squareKilometers;
-        },
-        _getArea: function (points) {
-            var earthRadius = 6378137;
-            var area = 0;
-            var len = points.length;
-            var x1 = points[len - 1].lng;
-            var y1 = points[len - 1].lat;
-            for (var i = 0; i < len; i++) {
-                var x2 = points[i].lng;
-                var y2 = points[i].lat;
-                area += this.toRadians(x2 - x1) * (2 + Math.sin(this.toRadians(y1)) + Math.sin(this.toRadians(y2)));
-                x1 = x2;
-                y1 = y2;
-            }
-            return Math.abs((area * earthRadius * earthRadius) / 2.0);
-        },
-        _numberFormat: function (number, decimals = 2) {
-            var thousandsSep = ",";
-            var sign = number < 0 ? "-" : "";
-            var num = Math.abs(+number || 0);
-            var intPart = parseInt(num.toFixed(decimals), 10) + "";
-            var j = intPart.length > 3 ? intPart.length % 3 : 0;
+      this._addNewVertex(latlng);
 
-            return [
-                sign,
-                j ? intPart.substr(0, j) + thousandsSep : "",
-                intPart.substr(j).replace(/(\d{3})(?=\d)/g, "$1" + thousandsSep),
-                decimals
-                    ? "." +
-                      Math.abs(num - intPart)
-                          .toFixed(decimals)
-                          .slice(2)
-                    : "",
-            ].join("");
-        },
-    });
+      if (this._measureBoundary) {
+        this._measureBoundary.bringToFront();
+      }
+      this._measureVertexes.bringToFront();
+    }
 
-    L.measureAction = function (map, options) {
-        return new L.MeasureAction(map, options);
-    };
-}, window);
+    this._updateResults();
+    this._updateMeasureStartedWithPoints();
+  },
+  // handle map mouse out during ongoing measure
+  // remove floating cursor vertex from map
+  _handleMapMouseOut: function() {
+    if (this._measureDrag) {
+      this._layer.removeLayer(this._measureDrag);
+      this._measureDrag = null;
+    }
+  },
+  // add various measure graphics to map - vertex, area, boundary
+  _addNewVertex: function(latlng) {
+    L.circleMarker(latlng, this._symbols.getSymbol('measureVertexActive')).addTo(
+      this._measureVertexes
+    );
+  },
+  _addMeasureArea: function(latlngs) {
+    if (latlngs.length < 3) {
+      if (this._measureArea) {
+        this._layer.removeLayer(this._measureArea);
+        this._measureArea = null;
+      }
+      return;
+    }
+    if (!this._measureArea) {
+      this._measureArea = L.polygon(latlngs, this._symbols.getSymbol('measureArea')).addTo(
+        this._layer
+      );
+    } else {
+      this._measureArea.setLatLngs(latlngs);
+    }
+  },
+  _addMeasureBoundary: function(latlngs) {
+    if (latlngs.length < 2) {
+      if (this._measureBoundary) {
+        this._layer.removeLayer(this._measureBoundary);
+        this._measureBoundary = null;
+      }
+      return;
+    }
+    if (!this._measureBoundary) {
+      this._measureBoundary = L.polyline(latlngs, this._symbols.getSymbol('measureBoundary')).addTo(
+        this._layer
+      );
+    } else {
+      this._measureBoundary.setLatLngs(latlngs);
+    }
+  }
+});
+
+L.Map.mergeOptions({
+  measureControl: false
+});
+
+L.Map.addInitHook(function() {
+  if (this.options.measureControl) {
+    this.measureControl = new L.Control.Measure().addTo(this);
+  }
+});
+
+L.control.measure = function(options) {
+  return new L.Control.Measure(options);
+};
